Add files via upload
TICKER is the stock ticker for the data. EST_DATE is the date of the analyst EPS estimate. FPI is the number of years ahead estimate. MEANEAST is the mean estimate of analysts' forecast of EPS. ACT_DATE	is the date of actual the actual EPS release. ACTUAL is the actual EPS value released.